Change submodule name to reflect refactor
Since the same class handles both adeft and gilda disambiguation, this
shouldn't be called adeft any longer
